Extract RootGauges repository
